first implementation of doppler shift (not working properly)
Integer colors!

+++ Programs/Skeleton/Skeleton/Color.h
@@ -14,28 +14,28 @@ static vec3 waveLengthToRGB(double Wavelength) {
     double Red, Green, Blue;
 
     if ((Wavelength >= 380) && (Wavelength < 440)) {
-        Red = -(Wavelength - 440) / (440 - 380);
+        Red = -(double)(Wavelength - 440) / (double)(440 - 380);
         Green = 0.0;
         Blue = 1.0;
     }
     else if ((Wavelength >= 440) && (Wavelength < 490)) {
         Red = 0.0;
-        Green = (Wavelength - 440) / (490 - 440);
+        Green = (double)(Wavelength - 440) / (double)(490 - 440);
         Blue = 1.0;
     }
     else if ((Wavelength >= 490) && (Wavelength < 510)) {
         Red = 0.0;
         Green = 1.0;
-        Blue = -(Wavelength - 510) / (510 - 490);
+        Blue = -(double)(Wavelength - 510) / (double)(510 - 490);
     }
     else if ((Wavelength >= 510) && (Wavelength < 580)) {
-        Red = (Wavelength - 510) / (580 - 510);
+        Red = (double)(Wavelength - 510) / (double)(580 - 510);
         Green = 1.0;
         Blue = 0.0;
     }
     else if ((Wavelength >= 580) && (Wavelength < 645)) {
         Red = 1.0;
-        Green = -(Wavelength - 645) / (645 - 580);
+        Green = -(double)(Wavelength - 645) / (double)(645 - 580);
         Blue = 0.0;
     }
     else if ((Wavelength >= 645) && (Wavelength < 781)) {
@@ -52,13 +52,13 @@ static vec3 waveLengthToRGB(double Wavelength) {
     // Let the intensity fall off near the vision limits
 
     if ((Wavelength >= 380) && (Wavelength < 420)) {
-        factor = 0.3 + 0.7 * (Wavelength - 380) / (420 - 380);
+        factor = 0.3 + 0.7 * (double)(Wavelength - 380) / (double)(420 - 380);
     }
     else if ((Wavelength >= 420) && (Wavelength < 701)) {
         factor = 1.0;
     }
     else if ((Wavelength >= 701) && (Wavelength < 781)) {
-        factor = 0.3 + 0.7 * (780 - Wavelength) / (780 - 700);
+        factor = 0.3 + 0.7 * (780 - Wavelength) / (double)(780 - 700);
     }
     else {
         factor = 0.0;
